Add support for migrations being programs
This is so we can run migrations as a separate process, and fail if
there is any stderr output. We'd like to be able to raise an exception
if there is a warning running Perl code, but doing so is not safe (see
https://perldoc.perl.org/warnings.html#Fatal-Warnings). Instead of doing
that, this will allow us to recognize that there was a warning, and
throw an error.